Refactoring app by omilovskiy's review
