--- src/main/java/com/cursor/hw_16_heroku/repo/LibraryRepo.java
@@ -1,14 +1,0 @@
-package com.cursor.hw_16_heroku.repo;
-
-import com.cursor.hw_16_heroku.model.Book;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface LibraryRepo extends CrudRepository<Book, Long> {
-
-    List<Book> findByGenre(String genre);
-
-    List<Book> findByAuthor(long authorId);
-
-}
