+++ src/main/resources/application.properties
@@ -8,4 +8,4 @@ spring.h2.console.path=/h2
 
 spring.jpa.show-sql=true
 
-server.port=8090
+server.port=8080
